requirements.txt
coverage==7.2.7
django-debug-toolbar==3.2.4
django==4.1
djangorestframework==3.14.0
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
flake8==5.0.4
pep8-naming==0.13.2
pytest==7.3.1
pytest-cov==4.1.0
